# openapi.yaml
...
openapi: 3.1.0
info:
  title: photo-portfolio-store
  version: 0.1.0
  description: |
    Photographer portfolio + print storefront API.

    Bounded contexts: auth, events, media, search, catalog, commerce,
    compliance, admin. Each operation declares its context via tag.
servers:
  - url: http://localhost:4000
    description: local
...
security: []
paths:
...
  /v1/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events visible to the caller
      x-status: stub
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/EventStatus' }
      responses:
        '200':
          description: Paginated events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      x-status: stub
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/events/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getEvent
      tags: [events]
      summary: Retrieve a single event
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Event resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Partially update an event
      x-status: stub
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPatch'
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Soft-delete (archive) an event
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Event archived.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/events/{id}/publish:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      operationId: publishEvent
      tags: [events]
      summary: Transition an event from draft to published
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Published event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/events/{id}/members:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      operationId: addEventMember
      tags: [events]
      summary: Add a member to an event
      x-status: stub
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventMemberCreate'
      responses:
        '201':
          description: Member added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMember'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/events/{id}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
      - in: path
        name: userId
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
    patch:
      operationId: updateEventMember
      tags: [events]
      summary: Update an event member's role or split
      x-status: stub
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventMemberPatch'
      responses:
        '200':
          description: Updated member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMember'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
    delete:
      operationId: removeEventMember
      tags: [events]
      summary: Remove an event member
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Member removed. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/events/{id}/ftp-credentials:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      operationId: rotateEventFtpCredentials
      tags: [events]
      summary: Issue (rotate) FTP credentials for the event
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: Newly issued credentials. The password is returned ONCE.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FtpCredentialIssued'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
        description: Opaque cursor returned by a prior page.
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict with current resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unprocessable:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            trace_id:
              type: string
...
    EventStatus:
      type: string
      enum: [draft, published, archived]
    EventMemberRole:
      type: string
      enum: [organizer, photographer, assistant]
    Event:
      type: object
      required:
        - id
        - orgId
        - name
        - slug
        - eventDate
        - timezone
        - status
        - allowFaceSearch
        - retentionDays
        - currency
        - createdAt
        - updatedAt
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        orgId: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        eventDate: { type: string, format: date }
        location: { type: string, nullable: true }
        timezone: { type: string }
        status: { $ref: '#/components/schemas/EventStatus' }
        allowFaceSearch: { type: boolean }
        retentionDays: { type: integer, minimum: 1 }
        currency: { type: string }
        coverPhotoId:
          oneOf:
            - { $ref: '#/components/schemas/Uuid' }
            - { type: 'null' }
        publishedAt: { type: string, format: date-time, nullable: true }
        archivedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EventCreate:
      type: object
      required: [orgId, name, slug, eventDate]
      properties:
        orgId: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        eventDate: { type: string, format: date }
        location: { type: string }
        timezone: { type: string, default: UTC }
        currency: { type: string, default: USD }
        allowFaceSearch: { type: boolean, default: true }
        retentionDays: { type: integer, default: 30, minimum: 1 }
    EventPatch:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        eventDate: { type: string, format: date }
        location: { type: string, nullable: true }
        timezone: { type: string }
        currency: { type: string }
        allowFaceSearch: { type: boolean }
        retentionDays: { type: integer, minimum: 1 }
        coverPhotoId: { $ref: '#/components/schemas/Uuid' }
    EventMember:
      type: object
      required: [eventId, userId, role, splitPct, createdAt]
      properties:
        eventId: { $ref: '#/components/schemas/Uuid' }
        userId: { $ref: '#/components/schemas/Uuid' }
        role: { $ref: '#/components/schemas/EventMemberRole' }
        splitPct:
          type: string
          description: Decimal as string (e.g. "33.33").
        createdAt: { type: string, format: date-time }
    EventMemberCreate:
      type: object
      required: [userId, role]
      properties:
        userId: { $ref: '#/components/schemas/Uuid' }
        role: { $ref: '#/components/schemas/EventMemberRole' }
        splitPct: { type: string, default: "0.00" }
    EventMemberPatch:
      type: object
      properties:
        role: { $ref: '#/components/schemas/EventMemberRole' }
        splitPct: { type: string }
    FtpCredentialIssued:
      type: object
      required: [id, eventId, username, password, expiresAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        eventId: { $ref: '#/components/schemas/Uuid' }
        username: { type: string }
        password:
          type: string
          description: Returned ONCE on rotation; never retrievable afterward.
        expiresAt: { type: string, format: date-time }
...
    PaginatedEvents:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Event' }
        nextCursor: { type: string, nullable: true }
...